requirements.txt
aiohttp
APScheduler
async-timeout
chardet
discord.py
emoji
multidict
pytz
six
tzlocal
websockets
pymongo
dnspython
